PWM on pin 4 (TIMER1B).
This fixes issue #36 (for Arduino 1.5.x at least). It uses
initVariant() to put Timer1 into PWM mode and a hacky #define alias
(TCCR1A for GTCCR) so that analogWrite() will set the right bit in the
right register to connect the timer PWM output to the pin. Also, the
pre-scale factor and PWM frequency are probably wrong (or, at least,
not the same as on other AVRs).